+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_standard/marines2/config.hpp
@@ -175,7 +175,7 @@ class 34thPRC_ArmourStandard_EOD_Helmet_Urban
 		label="Helmet Type";
 		alwaysSelectable=1;
 		changeingame=0;
-		values[]={"eod", "gungnir", "jfo", "cqc", "commando", "grenadier", "scout"};
+		values[]={"eod", "gungnir", "jfo", "cqc", "commando", "grenadier", "scout", "pilot"};
 		class eod
 		{
 			label="EOD";
@@ -203,6 +203,10 @@ class 34thPRC_ArmourStandard_EOD_Helmet_Urban
 		class scout
 		{
 			label="Scout";
+		};
+		class pilot
+		{
+			label="Pilot";
 		};
 	};
 	class camo